backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-groq
langgraph
python-dotenv
sqlalchemy
cors
